Add files via upload
Here,we have created a simple program in which we have several textboxes to note the name of tasks and corresponding checkboxes to check.If the tasks are completed by the user,the user will tick the tasks. It also consists of a button which completes all the tasks and closes the gui window.